examples/python-client/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
phlow-auth>=0.1.0